# docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ChangePasswordRequest" : {
        "properties" : {
          "currentPassword" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          },
          "newPassword" : {
            "type" : "string"
          }
        },
        "required" : [ "currentPassword", "newPassword" ],
        "type" : "object"
      },
...
      "UpdateProfileRequest" : {
        "properties" : {
          "fullName" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "phone" : {
            "pattern" : "^\\+[1-9]\\d{7,14}$",
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "fullName" ],
        "type" : "object"
      },
      "UserResponse" : {
        "properties" : {
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "fullName" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "phone" : {
            "type" : [ "string", "null" ]
          },
          "role" : {
            "enum" : [ "PASSENGER", "DRIVER", "ADMIN" ],
            "type" : "string"
          },
          "status" : {
            "enum" : [ "ACTIVE", "SUSPENDED" ],
            "type" : "string"
          }
        },
        "required" : [ "createdAt", "email", "fullName", "id", "role", "status" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Real-time ride-hailing platform API. Conventions: docs/api.md",
    "title" : "RideFlow API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/users/me" : {
      "get" : {
        "operationId" : "me",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Current user's profile",
        "tags" : [ "Users" ]
      },
      "patch" : {
        "operationId" : "update",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Update name and phone",
        "tags" : [ "Users" ]
      }
    },
    "/api/users/me/password" : {
      "put" : {
        "operationId" : "changePassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "summary" : "Change password; signs out all sessions",
        "tags" : [ "Users" ]
      }
    }
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "url" : "/"
  } ]
}